Changes the driver to work with Ubuntu 18.04 4-level PT structure

+++ softrmc_daemon/son_msg.cpp
@@ -44,8 +44,8 @@ using namespace std;
 
 void write_buf(char* buf, size_t len)
 {
-    for(int i = 0; i < len;i++) {
-        printf("Buffer[%d] = %c\n",i,buf[i]);
+    for(size_t i = 0; i < len;i++) {
+        printf("Buffer[%lu] = %c\n",i,buf[i]);
     }
 }
